+++ spec/spec_helper.rb
@@ -79,6 +79,7 @@ def inspect_source_file(cop, source)
 end
 
 def inspect_source(cop, source, file = nil)
+  Rubocop::Formatter::DisabledConfigFormatter.config_to_allow_offences = {}
   processed_source = parse_source(source, file)
   fail 'Error parsing example code' unless processed_source.valid_syntax?
   _investigate(cop, processed_source)